feature/variable-convention-update
This is a BREAKING change from previous versions. This change modifies
variable defaults to individual variables instead of within nested
blocks. This allows for overriding of defaults without needing to
specify the entire block. Without this change, attempting to modify a
single variable required the entire block to be defined within the role
vars.