feature/serial-over-usb-config
Modified kvmd override config to include a section for serial-over-usb
configuration. This is only applicable to pikvm version 2+ so a list of
version requirements were added to the vars. Additionally, modified the
readme.md to reflect the changes. All serial-over-usb services are
idempotent with these changes and should not affect existing
deployments. Lastly, modified the validation to ensure the needed files
for modification exist before starting.